Moved homescreen class to its own file in the screens folder
